Open the error logs only on failed downloads (#4)
I removed line buffering while I was at it. Within the process, there
is no need anymore to salvage as much log output as possible in the
face of an exception. Also, in the very rare event that two parallel
harvesters are writing to the same failure log at the same time, block
buffering will improve the odds that the backtraces are still readable.